requirements.txt
click==8.1.8
Flask==2.0.1
Flask-JWT-Extended==4.3.1
Flask-MySQLdb==0.2.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mysql==0.0.3
mysqlclient==2.2.7
PyJWT==2.10.1
PyMySQL==1.1.1
python-dotenv==0.19.0
Werkzeug==2.0.1
